requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Cython==0.29.33
dataclasses==0.6
differentiate==1.2.2
dirutility==0.7.20
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
looptools==1.2.4
mysql-connector==2.2.9
mysql-toolkit==1.9.1
numpy==1.21.6
packaging==23.0
paramiko==3.0.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
python-decouple==3.7
requests==2.28.2
rsa==4.9
SQLAlchemy==2.0.4
sqlparse==0.4.3
Telethon==1.27.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yadisk==1.2.19
zipp==3.15.0
